# openapi/baseline.json
{
  "components": {
    "schemas": {
...
      "Anomaly": {
        "properties": {
          "description": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CompareReportRequest": {
        "properties": {
          "activeFilters": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "fileIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "fileLabels": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "forceDirectedImage": {
            "type": "string"
          },
          "hierarchicalImage": {
            "type": "string"
          },
          "nodeLimitNote": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Correlation": {
        "properties": {
          "explanation": {
            "type": "string"
          },
          "externalEvent": {
            "type": "string"
          },
          "networkChange": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FileMetadataDto": {
        "properties": {
          "duration": {
            "format": "int64",
            "type": "integer"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "fileId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "type": "integer"
          },
          "packetCount": {
            "format": "int32",
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "FileUploadResponse": {
        "properties": {
          "fileId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "storageLocation": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "GenerateInsightRequest": {
        "properties": {
          "audience": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NarrativeSection": {
        "properties": {
          "content": {
            "type": "string"
          },
          "relatedData": {
            "$ref": "#/components/schemas/RelatedData"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "enum": [
              "summary",
              "detail",
              "anomaly",
              "conclusion"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NetworkInsightDto": {
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/Anomaly"
            },
            "type": "array"
          },
          "audience": {
            "type": "string"
          },
          "correlations": {
            "items": {
              "$ref": "#/components/schemas/Correlation"
            },
            "type": "array"
          },
          "errorMessage": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "modelUsed": {
            "type": "string"
          },
          "narrativeSections": {
            "items": {
              "$ref": "#/components/schemas/NarrativeSection"
            },
            "type": "array"
          },
          "networkId": {
            "format": "uuid",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PagedResponseFileMetadataDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FileMetadataDto"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "RelatedData": {
        "properties": {
          "conversations": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "packets": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ReportRequest": {
        "properties": {
          "activeFilters": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "forceDirectedImage": {
            "type": "string"
          },
          "hierarchicalImage": {
            "type": "string"
          },
          "nodeLimitNote": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "StepExplanation": {
        "properties": {
          "explanation": {
            "type": "string"
          },
          "stepIndex": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "StreamingResponseBody": {
        "type": "object"
      },
...
      "SubnetDefinitionDto": {
        "properties": {
          "cidr": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "densityScore": {
            "format": "double",
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "hostCount": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "snapshotsSeen": {
            "format": "int32",
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "totalSnapshots": {
            "format": "int32",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TracerExplainResponse": {
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "explanations": {
            "items": {
              "$ref": "#/components/schemas/StepExplanation"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TracerPeer": {
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "packetCount": {
            "format": "int64",
            "type": "integer"
          },
          "protocol": {
            "type": "string"
          },
          "responded": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "TracerPeersResponse": {
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "hostIp": {
            "type": "string"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/TracerPeer"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TracerStep": {
        "properties": {
          "direction": {
            "type": "string"
          },
          "info": {
            "type": "string"
          },
          "packetNumber": {
            "format": "int64",
            "type": "integer"
          },
          "payloadHex": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "stepIndex": {
            "format": "int32",
            "type": "integer"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TracerStepsResponse": {
        "properties": {
          "appName": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "dstIp": {
            "type": "string"
          },
          "dstPort": {
            "format": "int32",
            "type": "integer"
          },
          "protocol": {
            "type": "string"
          },
          "srcIp": {
            "type": "string"
          },
          "srcPort": {
            "format": "int32",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TracerStep"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "UpsertSubnetRequest": {
        "properties": {
          "cidr": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/files": {
      "get": {
        "description": "Get all uploaded files with pagination",
        "operationId": "getAllFiles",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "uploadedAt,desc",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "enum": [
                "ANALYSIS",
                "MONITOR"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseFileMetadataDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all files",
        "tags": [
          "File Management"
        ]
      },
      "post": {
        "description": "Upload a PCAP file for analysis",
        "operationId": "uploadFile",
        "parameters": [
          {
            "in": "query",
            "name": "enableNdpi",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableSuricata",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableFileExtraction",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "ANALYSIS",
              "enum": [
                "ANALYSIS",
                "MONITOR"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Upload PCAP file",
        "tags": [
          "File Management"
        ]
      }
    },
    "/api/v1/files/compare/report": {
      "post": {
        "operationId": "downloadCompareReport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate and download a PDF compare-topology report for multiple PCAP files",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/api/v1/files/{fileId}/extractions/{extractionId}/download": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "extractionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download an extracted file",
        "tags": [
          "Extracted Files"
        ]
      }
    },
...
    "/api/v1/files/{fileId}/report": {
      "post": {
        "operationId": "downloadReport",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate and download a PDF analysis report for a PCAP file",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/api/v1/monitor/networks/{networkId}/insights": {
      "post": {
        "operationId": "generate_1",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateInsightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkInsightDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate a new insight for a network",
        "tags": [
          "Network Insights"
        ]
      }
    },
...
    "/api/v1/monitor/networks/{networkId}/snapshots/{snapshotId}/insights": {
      "post": {
        "operationId": "generate",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "snapshotId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateInsightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkInsightDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate a new insight for a snapshot",
        "tags": [
          "Snapshot Insights"
        ]
      }
    },
...
    "/api/v1/subnets": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubnetDefinitionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List defined subnets",
        "tags": [
          "Subnets"
        ]
      },
      "post": {
        "operationId": "upsert_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSubnetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubnetDefinitionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create or update a subnet definition",
        "tags": [
          "Subnets"
        ]
      }
    },
    "/api/v1/subnets/detect": {
      "get": {
        "operationId": "detect",
        "parameters": [
          {
            "in": "query",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubnetDefinitionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Auto-detect subnets from a capture file",
        "tags": [
          "Subnets"
        ]
      }
    },
    "/api/v1/subnets/detect/network": {
      "get": {
        "operationId": "detectFromNetwork",
        "parameters": [
          {
            "in": "query",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubnetDefinitionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Auto-detect subnets across a monitored network",
        "tags": [
          "Subnets"
        ]
      }
    },
    "/api/v1/subnets/detected": {
      "post": {
        "operationId": "saveDetected",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSubnetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubnetDefinitionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Persist an auto-detected subnet",
        "tags": [
          "Subnets"
        ]
      }
    },
    "/api/v1/subnets/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete a subnet definition",
        "tags": [
          "Subnets"
        ]
      }
    },
...
    "/api/v1/tracer/{conversationId}/explanations": {
      "post": {
        "description": "Generates a plain-English explanation for each packet in the conversation using the configured LLM. Results are cached per conversation.",
        "operationId": "explainSteps",
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TracerExplainResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Generate AI explanations",
        "tags": [
          "Conversation Tracer"
        ]
      }
    },
    "/api/v1/tracer/{conversationId}/peers": {
      "get": {
        "description": "Returns every peer the traced host exchanged packets with, each flagged as responding or silent — for scan-style visualisation (e.g. ARP scans).",
        "operationId": "getPeers",
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TracerPeersResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get traced host's peers with response status",
        "tags": [
          "Conversation Tracer"
        ]
      }
    },
    "/api/v1/tracer/{conversationId}/steps": {
      "get": {
        "description": "Returns ordered packet steps for a conversation, ready for step-through visualisation.",
        "operationId": "getSteps",
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TracerStepsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get conversation steps",
        "tags": [
          "Conversation Tracer"
        ]
      }
    }
  },
...
}